Update to 1.21.8 + dump

+++ src/main/java/uk/protonull/civvoxelmap/mixins/elevation/ObscureElevationMixin.java
@@ -21,7 +21,7 @@ public abstract class ObscureElevationMixin {
     @Shadow(
         remap = false
     )
-    public int y;
+    public double y;
 
     @Inject(
         method = "updateLocation",
